email-notifier/requirements.txt
kafka-python==2.0.2
redis==4.6.0
python-dateutil==2.8.2 